requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
colorama==0.4.6
coverage==7.13.5
greenlet==3.3.2
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.1.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
